A bunch of changes for 64-bit system compatibility: * Add a SerializeAction process method for a u_int64_t * Change all size_t parameters in the various serialize functions to   explicitly take a u_int32_t so that the size is constant across   systems * Change some serialized values that used to be a size_t to a   u_int32_t to account for the above * Fix unit tests to work on 64-bit systems

+++ serialize/TypeShims.h
@@ -129,7 +129,7 @@ public:
     // virtual from SerializableObject
     void serialize(SerializeAction* a)
     {
-        size_t len = 0;
+        u_int32_t len = 0;
         a->process(name_.c_str(), 
 		   reinterpret_cast<u_char**>(&str_), &len,
 		   Serialize::NULL_TERMINATED | Serialize::ALLOC_MEM);
@@ -177,12 +177,12 @@ public:
  
     const char* value() const { return buf_; }
     char*       take_buf()    { own_buf_ = false; return buf_; }
-    size_t      size()  const { return size_; }
+    u_int32_t   size()  const { return size_; }
 
 private:
-    char*  buf_;
-    size_t size_;    
-    bool   own_buf_;
+    char*     buf_;
+    u_int32_t size_;    
+    bool      own_buf_;
 };
 
 //----------------------------------------------------------------------------
